Big Push - (mostly complete)
Should have done more pushes hehe.
Created a proxy API for google search. Styled website using styled-components. Implemented infinite scroll within bounds of the API. Considered implementing cashing with redis but decided against it, the rate limit should be enough for this demonstration and the ammount of different searches that can be made makes it infeasable for a project like this. Used hooks to add the necessary functionality. Implemented auto search with a delay, no search button required. Implemented a primitive form of error handling. Used environment variables to hide the API key. Switched out to my own favicon. Made the default search my old home town. Added a scroll-up button to take the user back to the top of the page. And more that I'm probably forgetting.

TODO:
Make it look less bland.
Add comments to explain the code.